Break the media class into different classes.
