+++ includes/Media/Video/Optimization.php
@@ -1,15 +1,15 @@
 <?php
 /**
- * Class Video_Captions
+ * Class Optimization
  *
  * @package   Google\Web_Stories
- * @copyright 2020 Google LLC
+ * @copyright 2021 Google LLC
  * @license   https://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/google/web-stories-wp
  */
 
 /**
- * Copyright 2020 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,40 +24,46 @@
  * limitations under the License.
  */
 
-namespace Google\Web_Stories\Media;
+namespace Google\Web_Stories\Media\Video;
 
 use Google\Web_Stories\Service_Base;
 
 /**
- * Class Video_Captions
+ * Class Optimization
  *
- * @package Google\Web_Stories
+ * @package Google\Web_Stories\Media\Video
  */
-class Video_Captions extends Service_Base {
+class Optimization extends Service_Base {
+
 	/**
-	 * Initializes the File_Type logic.
+	 * The optimized video id post meta key.
 	 *
-	 * @since 1.7.0
+	 * @var string
+	 */
+	const OPTIMIZED_ID_POST_META_KEY = 'web_stories_optimized_id';
+
+	/**
+	 * Init.
+	 *
+	 * @since 1.10.0
 	 *
 	 * @return void
 	 */
 	public function register() {
-		add_filter( 'site_option_upload_filetypes', [ $this, 'filter_list_of_allowed_filetypes' ] );
+		register_meta(
+			'post',
+			self::OPTIMIZED_ID_POST_META_KEY,
+			[
+				'sanitize_callback' => 'absint',
+				'type'              => 'integer',
+				'description'       => __( 'ID of optimized video.', 'web-stories' ),
+				'show_in_rest'      => true,
+				'default'           => 0,
+				'single'            => true,
+				'object_subtype'    => 'attachment',
+			]
+		);
 	}
 
-	/**
-	 * Add VTT file type to allow file in multisite.
-	 *
-	 * @param string $value List of allowed file types.
-	 * @return string List of allowed file types.
-	 */
-	public function filter_list_of_allowed_filetypes( $value ): string {
-		$filetypes = explode( ' ', $value );
-		if ( ! in_array( 'vtt', $filetypes, true ) ) {
-			$filetypes[] = 'vtt';
-			$value       = implode( ' ', $filetypes );
-		}
 
-		return $value;
-	}
 }
