+++ includes/Plugin.php
@@ -91,10 +91,15 @@ class Plugin extends ServiceBasedPlugin {
 		'integrations.themes_support'  => Integrations\Core_Themes_Support::class,
 		'imgareaselect_patch'          => Admin\ImgAreaSelect_Patch::class,
 		'kses'                         => KSES::class,
-		'media'                        => Media\Media::class,
 		'page_template_post_type'      => Page_Template_Post_Type::class,
 		'plugin_row_meta'              => Admin\PluginRowMeta::class,
 		'plugin_action_links'          => Admin\PluginActionLinks::class,
+		'media.image_sizes'            => Media\Image_Sizes::class,
+		'media.media_source'           => Media\Media_Source_Taxonomy::class,
+		'media.video.captions'         => Media\Video\Captions::class,
+		'media.video.muting'           => Media\Video\Muting::class,
+		'media.video.optimization'     => Media\Video\Optimization::class,
+		'media.video.poster'           => Media\Video\Poster::class,
 		'meta_boxes'                   => Admin\Meta_Boxes::class,
 		'settings'                     => Settings::class,
 		'site_health'                  => Admin\Site_Health::class,
@@ -122,8 +127,6 @@ class Plugin extends ServiceBasedPlugin {
 		'rest.template_lock'           => REST_API\Template_Lock_Controller::class,
 		'user_preferences'             => User\Preferences::class,
 		'web_stories_block'            => Block\Web_Stories_Block::class,
-		'video_captions'               => Media\Video_Captions::class,
-		'video_muting'                 => Media\Video_Muting::class,
 	];
 
 	/**
