+++ includes/REST_API/Stories_Controller.php
@@ -27,7 +27,7 @@
 namespace Google\Web_Stories\REST_API;
 
 use Google\Web_Stories\Demo_Content;
-use Google\Web_Stories\Media\Media;
+use Google\Web_Stories\Media\Image_Sizes;
 use Google\Web_Stories\Settings;
 use Google\Web_Stories\Story_Post_Type;
 use Google\Web_Stories\Traits\Publisher;
@@ -92,7 +92,7 @@ class Stories_Controller extends Stories_Base_Controller {
 		}
 
 		if ( rest_is_field_included( 'featured_media_url', $fields ) ) {
-			$image                      = get_the_post_thumbnail_url( $post, Media::POSTER_PORTRAIT_IMAGE_SIZE );
+			$image                      = get_the_post_thumbnail_url( $post, Image_Sizes::POSTER_PORTRAIT_IMAGE_SIZE );
 			$data['featured_media_url'] = ! empty( $image ) ? $image : $schema['properties']['featured_media_url']['default'];
 		}
