+++ includes/uninstall.php
@@ -27,8 +27,10 @@
 namespace Google\Web_Stories;
 
 use Google\Web_Stories\User\Preferences;
-use Google\Web_Stories\Media\Media;
-use Google\Web_Stories\Media\Video_Muting;
+use Google\Web_Stories\Media\Media_Source_Taxonomy;
+use Google\Web_Stories\Media\Video\Optimization;
+use Google\Web_Stories\Media\Video\Muting;
+use Google\Web_Stories\Media\Video\Poster;
 
 /**
  * Deletes options and transients.
@@ -117,11 +119,11 @@ function delete_site_options() {
  * @return void
  */
 function delete_stories_post_meta() {
-	delete_post_meta_by_key( Media::POSTER_POST_META_KEY );
-	delete_post_meta_by_key( Media::POSTER_ID_POST_META_KEY );
-	delete_post_meta_by_key( Media::OPTIMIZED_ID_POST_META_KEY );
-	delete_post_meta_by_key( Video_Muting::MUTED_ID_POST_META_KEY );
-	delete_post_meta_by_key( Video_Muting::IS_MUTED_POST_META_KEY );
+	delete_post_meta_by_key( Poster::POSTER_POST_META_KEY );
+	delete_post_meta_by_key( Poster::POSTER_ID_POST_META_KEY );
+	delete_post_meta_by_key( Optimization::OPTIMIZED_ID_POST_META_KEY );
+	delete_post_meta_by_key( Muting::MUTED_ID_POST_META_KEY );
+	delete_post_meta_by_key( Muting::IS_MUTED_POST_META_KEY );
 }
 
 /**
@@ -171,7 +173,7 @@ function delete_posts() {
  * @return void
  */
 function delete_terms() {
-	$taxonomy = Media::STORY_MEDIA_TAXONOMY;
+	$taxonomy = Media_Source_Taxonomy::TAXONOMY_SLUG;
 	$term_ids = get_terms(
 		[
 			'taxonomy'   => $taxonomy,
