+++ tests/phpunit/tests/Media/Video/Muting.php
@@ -15,20 +15,20 @@
  * limitations under the License.
  */
 
-namespace Google\Web_Stories\Tests\Media;
+namespace Google\Web_Stories\Tests\Media\Video;
 
 use Google\Web_Stories\Tests\Test_Case;
 use WP_REST_Request;
 
 /**
- * @coversDefaultClass \Google\Web_Stories\Media\Video_Muting
+ * @coversDefaultClass \Google\Web_Stories\Media\Video\Muting
  */
-class Video_Muting extends Test_Case {
+class Muting extends Test_Case {
 	/**
 	 * @covers ::register
 	 */
 	public function test_register() {
-		$video_muting = new \Google\Web_Stories\Media\Video_Muting();
+		$video_muting = new \Google\Web_Stories\Media\Video\Muting();
 		$video_muting->register();
 
 		$this->assertSame( 10, has_action( 'rest_api_init', [ $video_muting, 'rest_api_init' ] ) );
@@ -48,11 +48,11 @@ class Video_Muting extends Test_Case {
 	 * @covers ::register_meta
 	 */
 	public function test_register_meta() {
-		$video_muting = new \Google\Web_Stories\Media\Video_Muting();
+		$video_muting = new \Google\Web_Stories\Media\Video\Muting();
 		$this->call_private_method( $video_muting, 'register_meta' );
 
-		$this->assertTrue( registered_meta_key_exists( 'post', \Google\Web_Stories\Media\Video_Muting::IS_MUTED_POST_META_KEY, 'attachment' ) );
-		$this->assertTrue( registered_meta_key_exists( 'post', \Google\Web_Stories\Media\Video_Muting::MUTED_ID_POST_META_KEY, 'attachment' ) );
+		$this->assertTrue( registered_meta_key_exists( 'post', \Google\Web_Stories\Media\Video\Muting::IS_MUTED_POST_META_KEY, 'attachment' ) );
+		$this->assertTrue( registered_meta_key_exists( 'post', \Google\Web_Stories\Media\Video\Muting::MUTED_ID_POST_META_KEY, 'attachment' ) );
 	}
 
 	/**
@@ -80,7 +80,7 @@ class Video_Muting extends Test_Case {
 
 		set_post_thumbnail( $video_attachment_id, $poster_attachment_id );
 
-		$video_muting = new \Google\Web_Stories\Media\Video_Muting();
+		$video_muting = new \Google\Web_Stories\Media\Video\Muting();
 		$image        = $video_muting->wp_prepare_attachment_for_js(
 			[
 				'id'   => $poster_attachment_id,
@@ -96,9 +96,8 @@ class Video_Muting extends Test_Case {
 			]
 		);
 
-
-		$this->assertArrayNotHasKey( \Google\Web_Stories\Media\Video_Muting::IS_MUTED_KEY, $image );
-		$this->assertArrayHasKey( \Google\Web_Stories\Media\Video_Muting::IS_MUTED_KEY, $video );
+		$this->assertArrayNotHasKey( \Google\Web_Stories\Media\Video\Muting::IS_MUTED_KEY, $image );
+		$this->assertArrayHasKey( \Google\Web_Stories\Media\Video\Muting::IS_MUTED_KEY, $video );
 	}
 
 	/**
@@ -120,15 +119,15 @@ class Video_Muting extends Test_Case {
 		$response = rest_get_server()->dispatch( $request );
 		$data     = $response->get_data();
 
-		$this->assertArrayHasKey( \Google\Web_Stories\Media\Video_Muting::IS_MUTED_KEY, $data );
-		$this->assertNull( $data[ \Google\Web_Stories\Media\Video_Muting::IS_MUTED_KEY ] );
+		$this->assertArrayHasKey( \Google\Web_Stories\Media\Video\Muting::IS_MUTED_KEY, $data );
+		$this->assertNull( $data[ \Google\Web_Stories\Media\Video\Muting::IS_MUTED_KEY ] );
 	}
 
 	/**
 	 * @covers ::get_callback_is_muted
 	 */
 	public function test_get_callback_is_muted() {
-		$video_muting        = new \Google\Web_Stories\Media\Video_Muting();
+		$video_muting        = new \Google\Web_Stories\Media\Video\Muting();
 		$video_attachment_id = self::factory()->attachment->create_object(
 			[
 				'file'           => DIR_TESTDATA . '/uploads/test-video.mp4',
@@ -137,7 +136,7 @@ class Video_Muting extends Test_Case {
 				'post_title'     => 'Test Video',
 			]
 		);
-		update_post_meta( $video_attachment_id, \Google\Web_Stories\Media\Video_Muting::IS_MUTED_POST_META_KEY, '1' );
+		update_post_meta( $video_attachment_id, \Google\Web_Stories\Media\Video\Muting::IS_MUTED_POST_META_KEY, '1' );
 		$result = $video_muting->get_callback_is_muted( [ 'id' => $video_attachment_id ] );
 		$this->assertTrue( $result );
 	}
@@ -147,7 +146,7 @@ class Video_Muting extends Test_Case {
 	 */
 	public function test_update_callback_is_muted() {
 		wp_set_current_user( self::factory()->user->create( [ 'role' => 'administrator' ] ) );
-		$video_muting        = new \Google\Web_Stories\Media\Video_Muting();
+		$video_muting        = new \Google\Web_Stories\Media\Video\Muting();
 		$video_attachment_id = self::factory()->attachment->create_object(
 			[
 				'file'           => DIR_TESTDATA . '/uploads/test-video.mp4',
@@ -167,7 +166,7 @@ class Video_Muting extends Test_Case {
 	 * @covers ::update_callback_is_muted
 	 */
 	public function test_update_callback_is_muted_error() {
-		$video_muting        = new \Google\Web_Stories\Media\Video_Muting();
+		$video_muting        = new \Google\Web_Stories\Media\Video\Muting();
 		$video_attachment_id = self::factory()->attachment->create_object(
 			[
 				'file'           => DIR_TESTDATA . '/uploads/test-video.mp4',
